#239 move to `front_assets`
